Track your favorite Twitch streamers with this easy, sleek desktop widget!
For a quick how-to on using the program please watch the following
video.

https://www.youtube.com/watch?v=nzSv_KG0v_c

But what this program does it allows you to add streamers to a GUI and
have it either automatically or manually update when a streamer goes
live!

This widget sits nicely on your desktop and out of the way.